implemented a global handler for validation

+++ src/main/kotlin/com/sebqvcoding/springboot/note/controller/NotesController.kt
@@ -4,8 +4,11 @@ import com.sebqvcoding.springboot.note.controller.core.NoteRequest
 import com.sebqvcoding.springboot.note.controller.core.NoteResponse
 import com.sebqvcoding.springboot.note.database.mapper.NoteMapper
 import com.sebqvcoding.springboot.note.database.repository.NotesRepository
+import jakarta.validation.Valid
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
+import java.lang.IllegalArgumentException
 
 @RestController
 @RequestMapping("/notes")
@@ -15,21 +18,26 @@ class NotesController constructor(
 ) {
 
     @PostMapping
-    fun saveNote(@RequestBody body: NoteRequest): NoteResponse {
-        val savedNote = notesRepository.save(noteMapper.mapToNote(body))
+    fun saveNote(@RequestBody @Valid body: NoteRequest): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        val savedNote = notesRepository.save(noteMapper.mapToNote(body, ownerId))
         return noteMapper.mapToNotesResponse(savedNote)
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(name = "ownerId", required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return notesRepository.findByOwnerId(ownerId = ObjectId(ownerId))
             .map { noteMapper.mapToNotesResponse(it) }
     }
 
     @DeleteMapping(path = ["/{id}"])
-    fun deleteById(@PathVariable id:String){
-        notesRepository.deleteById(ObjectId(id))
+    fun deleteById(@PathVariable id: String) {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        val note = notesRepository.findById(ObjectId(id)).orElseThrow {
+            throw IllegalArgumentException("Note not found")
+        }
+        if (note.ownerId.toHexString() == ownerId)
+            notesRepository.deleteById(ObjectId(id))
     }
 }
